store/tikv: move ticlient_test to /tests
